CHET-107: Refactor AWS Region management to be separate to credentials management
