+++ src/main/java/com/atlassian/migration/datacenter/core/aws/auth/PlainTextCredentialsManager.java
@@ -1,6 +1,5 @@
 package com.atlassian.migration.datacenter.core.aws.auth;
 
-import com.atlassian.migration.datacenter.core.aws.GlobalInfrastructure;
 import com.atlassian.plugin.spring.scanner.annotation.imports.ComponentImport;
 import com.atlassian.sal.api.pluginsettings.PluginSettings;
 import com.atlassian.sal.api.pluginsettings.PluginSettingsFactory;
@@ -13,20 +12,17 @@ import org.springframework.stereotype.Component;
  * and should not be used in production.
  */
 @Component
-public class PlainTextCredentialsManager implements CredentialsFetcher, CredentialsStorer, RegionFetcher, RegionStorer {
+public class PlainTextCredentialsManager implements CredentialsFetcher, CredentialsStorer {
 
     private static final String AWS_CREDS_PLUGIN_STORAGE_KEY = "com.atlassian.migration.datacenter.core.aws.auth";
     private static final String ACCESS_KEY_ID_PLUGIN_STORAGE_SUFFIX = ".accessKeyId";
     private static final String SECRET_ACCESS_KEY_PLUGIN_STORAGE_SUFFIX = ".secretAccessKey";
-    private static final String REGION_PLUGIN_STORAGE_SUFFIX = ".region";
 
-    private final GlobalInfrastructure globalInfrastructure;
     private final PluginSettingsFactory pluginSettingsFactory;
 
     @Autowired
-    public PlainTextCredentialsManager(@ComponentImport PluginSettingsFactory pluginSettingsFactory, GlobalInfrastructure globalInfrastructure) {
+    public PlainTextCredentialsManager(@ComponentImport PluginSettingsFactory pluginSettingsFactory) {
         this.pluginSettingsFactory = pluginSettingsFactory;
-        this.globalInfrastructure = globalInfrastructure;
     }
 
     @Override
@@ -42,12 +38,6 @@ public class PlainTextCredentialsManager implements CredentialsFetcher, Credenti
     }
 
     @Override
-    public String getRegion() {
-        PluginSettings pluginSettings = pluginSettingsFactory.createGlobalSettings();
-        return (String) pluginSettings.get(AWS_CREDS_PLUGIN_STORAGE_KEY + REGION_PLUGIN_STORAGE_SUFFIX);
-    }
-
-    @Override
     public void storeAccessKeyId(String accessKeyId) {
         PluginSettings pluginSettings = pluginSettingsFactory.createGlobalSettings();
         pluginSettings.put(AWS_CREDS_PLUGIN_STORAGE_KEY + ACCESS_KEY_ID_PLUGIN_STORAGE_SUFFIX, accessKeyId);
@@ -57,22 +47,5 @@ public class PlainTextCredentialsManager implements CredentialsFetcher, Credenti
     public void storeSecretAccessKey(String secretAccessKey) {
         PluginSettings pluginSettings = pluginSettingsFactory.createGlobalSettings();
         pluginSettings.put(AWS_CREDS_PLUGIN_STORAGE_KEY + SECRET_ACCESS_KEY_PLUGIN_STORAGE_SUFFIX, secretAccessKey);
-    }
-
-    @Override
-    public void storeRegion(String region) throws InvalidAWSRegionException {
-        if(!isValidRegion(region)) {
-            throw new InvalidAWSRegionException();
-        }
-
-        PluginSettings pluginSettings = pluginSettingsFactory.createGlobalSettings();
-        pluginSettings.put(AWS_CREDS_PLUGIN_STORAGE_KEY + REGION_PLUGIN_STORAGE_SUFFIX, region);
-    }
-
-    private boolean isValidRegion(String testRegion) {
-        return globalInfrastructure.
-                getRegions()
-                .stream()
-                .anyMatch(region -> region.equals(testRegion));
     }
 }
